+++ app/constants/index.ts
@@ -1,2 +1,3 @@
 export const USDC_ADDRESS = '0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913';
+export const USDC_DECIMALS = 6;
 export const BASE_CHAIN_ID = 8453;
